pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lemlem"
version = "0.1.0"
description = "Minimal light wrapper for LLM calls with fallbacks"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [
  { name = "Evergreen Team" }
]
dependencies = [
  "openai>=1.0.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["lemlem*"]

